Ball moves

+++ src/hub.rs
@@ -79,7 +79,7 @@ impl<CLK, OEN, LT, A, B, C, R1, G1, B1, R2, G2, B2> HUBPort<CLK, OEN, LT, A, B, 
          self.latch.set_low().ok();
          self.output_enabled.set_low().ok();
      }
-     pub(crate) fn next_pixel(&mut self, pixel: u16) {
+     pub(crate) fn next_pixel(&mut self, pixel: u8) {
          self.clock.set_high().ok();
 
          if ((pixel >> 5) & 0b001) == 1 {
